Added a Readme file(pending update), a functioning HTML and two javascript files
The html document sets an outline for the page, and references google's jQuery, bootstrap v5.0 -bundle. The script.js does most of the work in setting up the page and user interactivity. The questions.js is a script holding an array of questions, and users can add their own questions by following my comments